vision/requirements.txt
# Vision Pipeline Dependencies
opencv-python==4.8.1.78
pdf2image==1.16.3
Pillow==10.1.0
numpy==1.26.2
fastapi==0.104.1
python-multipart==0.0.6
uvicorn==0.24.0
pydantic==2.5.0
